% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-constructors.R
\name{agg_abline}
\alias{agg_abline}
\title{Add a line between two points on your graph}
\usage{
agg_abline(x = NULL, y = NULL, x1, y1, x2, y2, colour = "black",
  panel, lwd = 1, lty = 1, color)
}
\arguments{
\item{x}{(Deprecated; use agg_vline instead) Draw a vertical line at x}

\item{y}{(Deprecated; use agg_hline instead) Draw a horizontal at y}

\item{x1}{The first x coordinate}

\item{y1}{The first y coordinate}

\item{x2}{The second x coordinate}

\item{y2}{The second y coordinate}

\item{colour}{The colour of the AB line (default black)}

\item{panel}{Which panel should the line be placed on? You can specify a
vector of panels (e.g. `panel = c("1","3")`) to apply the line to multiple
panels at once.}

\item{lwd}{(Optional, default 1) The line width}

\item{lty}{(Optional, default 1) The line type  (uses R line types)}

\item{color}{(Deprecated; use colour instead) The colour of the AB line
(default black)}
}
\description{
Add a line between two points on your graph
}
\examples{
arphitgg(data) + agg_abline(x = 2001, colour = RBA["Blue1"], panel = "1") +
  agg_abline(y = -0.5, colour = RBA["Red1"], panel = "1")

arphitgg(data) +
  agg_abline(x1 = 2000, y1 = -0.1, x2 = 2002, y2 = 0.5,
             colour = RBA["Blue1"], panel = "1")

}
\seealso{
\code{vignette("plotting-options", package = "arphit")} for a
detailed description of all the plotting options
}
