% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{agg_aes}
\alias{agg_aes}
\title{Define an aesthetic for a graph, or a graph layer.}
\usage{
agg_aes(x, y, group = NULL)
}
\arguments{
\item{x}{Which series is the x variable.}

\item{y}{Which series are you plotting on the y axis.}

\item{group}{If your data are in long form, which variable defines the groups.}
}
\description{
If specified as part of a agg_line or agg_col, fields left blank will be inherited from the parent.
}
\seealso{
\code{vignette("gg-interface", package = "arphit")} for a detailed description of
how to use the ggplot-like interface.
}
