% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-constructors.R
\name{agg_arrow}
\alias{agg_arrow}
\title{Add an arrow}
\usage{
agg_arrow(tail.x, tail.y, head.x, head.y, color, panel, lwd = 1)
}
\arguments{
\item{tail.x}{The x coordinate of the arrow tail}

\item{tail.y}{The y coordinate of the arrow tail}

\item{head.x}{The x coordinate of the arrow head}

\item{head.y}{The y coordinate of the arrow head}

\item{color}{The color of the arrow}

\item{panel}{Which panel should the arrow be placed on?}

\item{lwd}{(Optional, default 1) The linewidth of the arrow}
}
\description{
Add an arrow
}
\examples{
arphitgg(data) +
  agg_arrow(tail.x = 2002, tail.y = 0, head.x = 2003, head.y = 1,
            color = RBA["Blue1"], panel = "1")

}
\seealso{
\code{vignette("gg-interface", package = "arphit")} for a detailed description of
how to use the ggplot-like interface.
}
