% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-constructors.R
\name{agg_autolabel}
\alias{agg_autolabel}
\title{Add automatically placed label}
\usage{
agg_autolabel()
}
\description{
Add automatically placed label
}
\examples{
arphitgg(data) + agg_line() + agg_autolabel()

}
\seealso{
\code{vignette("plotting-options", package = "arphit")} for a detailed description of
all the plotting options
}
