% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-operators.R
\name{agg_draw}
\alias{agg_draw}
\title{Draw a defined graph}
\usage{
agg_draw(gg, filename = NULL)
}
\arguments{
\item{gg}{An arphitgg built graph.}

\item{filename}{(optional) If specified, save image to filename instead of displaying in R. Supports pdf, emf and png extensions.}
}
\description{
Draw a defined graph
}
\seealso{
\code{vignette("gg-interface", package = "arphit")} for a detailed description of
how to use the ggplot-like interface.
}
