% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-constructors.R
\name{agg_footnote}
\alias{agg_footnote}
\title{Add a footnote (or many footnotes)}
\usage{
agg_footnote(footnote)
}
\arguments{
\item{footnote}{A string, or vector of strings, to be added as footnotes.}
}
\description{
Add a footnote (or many footnotes)
}
\examples{
arphitgg(data) + agg_footnote("Here is a footnote")
arphitgg(data) + agg_footnote(c("Here is a footnote", "And a second one"))

}
\seealso{
\code{vignette("plotting-options", package = "arphit")} for a detailed description of
all the plotting options
}
