% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-constructors.R
\name{agg_label}
\alias{agg_label}
\title{Add a label}
\usage{
agg_label(text, x, y, panel, color = "black", size = 20)
}
\arguments{
\item{text}{The text to display on your plot}

\item{x}{The x coordinate of the center of your label}

\item{y}{The y coordinate of the center of your label}

\item{panel}{Which panel should the label be placed on?}

\item{color}{The color of your text}

\item{size}{Font size (default 20)}
}
\description{
Add a label
}
\examples{
arphitgg(data) + agg_label("Here is a footnote", RBA["Red3"], 2003, 0.2, "1")

}
\seealso{
\code{vignette("plotting-options", package = "arphit")} for a detailed description of
all the plotting options
}
