% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-constructors.R
\name{agg_line}
\alias{agg_line}
\title{Add a line layer to an arphit plot.}
\usage{
agg_line(aes = NULL, data = NULL, colour = NULL, color, pch = NULL,
  lty = NULL, lwd = NULL, pointsize = 1, panel = "1")
}
\arguments{
\item{aes}{The aesthetic that defines the layer. Will inherit (or parts thereof) if omitted.}

\item{data}{The data to be used. Will inherit from parent if missing.}

\item{colour}{A colour to be applied to all of the series, or (if your aesthetic has a group), a vector of colours that will be cycled through to consecutive group elements.}

\item{color}{(Deprecated; use colour instead) A colour to be applied to all of the series, or (if your aesthetic has a group), a vector of colours that will be cycled through to consecutive group elements.}

\item{pch}{A point marker to be applied to all series, or or (if your aesthetic has a group), a vector of pch values that will be cycled through to consecutive group elements. Any value accepted by R for pch can be used.}

\item{lty}{A line type to be applied to all series, or or (if your aesthetic has a group), a vector of lty values that will be cycled through to consecutive group elements. Any value accepted by R for lty can be used.}

\item{lwd}{A line width to be applied to all series, or or (if your aesthetic has a group), a vector of lwd values that will be cycled through to consecutive group elements. Any value accepted by R for lwd can be used.}

\item{pointsize}{Scale the size of the points? (default 1)}

\item{panel}{(default = "1") Which panel of the graph to place this layer on. You can specify a vector of panels (e.g. `panel = c("1","3")`) to apply the layer to multiple panels at once.}
}
\description{
Add a line layer to an arphit plot.
}
\examples{
data  <- data.frame(unemployment = rnorm(20), state = c(rep("A", 10), rep("B", 10)),
  date = seq.Date(from = as.Date("2017-01-10"), length.out = 10, by = "quarter"))
arphitgg(data) + agg_line(aes = agg_aes(x = date, y = unemployment, group = state), panel = "1")

}
\seealso{
\code{vignette("plotting-options", package = "arphit")} for a detailed description of
all the plotting options
}
