% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-constructors.R
\name{agg_point}
\alias{agg_point}
\title{Add a scatter layer to an arphit plot.}
\usage{
agg_point(aes = NULL, data = NULL, color = NULL, pointsize = 1,
  panel = "1")
}
\arguments{
\item{aes}{The aesthetic that defines the layer. Will inherit (or parts thereof) if omitted.}

\item{data}{The data to be used. Will inherit from parent if missing.}

\item{color}{A colour to be applied to all of the series, or (if your aesthetic has a group), a vector of colours that will be cycled through to consecutive group elements.}

\item{pointsize}{Scale the size of the points? (default 1)}

\item{panel}{(default = "1") Which panel of the graph to place this layer on.}
}
\description{
Add a scatter layer to an arphit plot.
}
\examples{
data  <- data.frame(x = rnorm(10), y = rnorm(10))
arphitgg(data) + agg_point(aes = agg_aes(x = x, y = y), panel = "1")

}
\seealso{
\code{vignette("plotting-options", package = "arphit")} for a detailed description of
all the plotting options
}
