% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot.R
\name{agg_qplot}
\alias{agg_qplot}
\title{Quick plot - for quickly creates a single-panel graph. Supports bar and line (and combinations of).}
\usage{
agg_qplot(data, series = NULL, x = NULL, bars = NULL,
  filename = NULL, title = NULL, subtitle = NULL, footnotes = c(),
  sources = c(), yunits = NULL, col = list(), pch = list(),
  lty = list(), lwd = list(), xlim = list(), ylim = list(),
  legend = FALSE, legend.ncol = NA, bar.stacked = TRUE)
}
\arguments{
\item{data}{Object containing the series you want to plot. Can be a data.frame, tibble or ts. Can also be a list of the above, with separate for each panel.}

\item{series}{A vector of series names specifying which subset of series
you want to plot.}

\item{x}{The x variable for your plot. `ts`, `xts` and `zoo` data use the dates in the time series.}

\item{bars}{(optional) Vector of string names indicating which series should be bars, rather than lines. Alternatively, if you set `bars = TRUE` all series will plot as bars.}

\item{filename}{(optional) If specified, save image to filename instead of displaying in R. Supports pdf, emf and png extensions.}

\item{title}{(optional) A string indicating the title for the entire chart. Passing NULL (or omitting the argument) will suppress printing of title.}

\item{subtitle}{(optional) A string indicating the subtitle for the entire chart. Passing NULL (or omitting the argument) will suppress printing of subtitle.}

\item{footnotes}{(optional) A vector strings, corresponding to the footnotes, in order.}

\item{sources}{(optional) A vector of strings, one entry for each source.}

\item{yunits}{(optional) A list of string -> string pairs indicating the units to be used on each panel (/axes, see notes to series for explanation on how right-hand-side axes are treated). Alternatively, providing just a string will assign that to all panels. If not supplied, a per cent sign will be used.}

\item{col}{(optional) A list of string -> misc pairs. The keys should be series names, and the values colours for each series (any colour accepted by R is fine.) You need not supply colours for all series. Default colours will be assigned  (cycling through) to series without assigned colours. Alternatively, you can supply a single value to apply to all series.}

\item{pch}{(optional) A list of string -> int pairs. The keys should be series names, and the values markers (pch values) for each series. Defaults to none (NA). Can be supplied as list, or a single value (which will be applied to all series).}

\item{lty}{(optional) A list of string -> int pairs. The keys should be series names, and the values line types (lty values) for each series. Defaults to solid (1). Can be supplied as list, or a single value (which will be applied to all series).}

\item{lwd}{(optional) A list of string -> numeric pairs. The keys should be series names, and the values line width, relative to default, for each series. Defaults to 1. Can be supplied as list, or a single value (which will be applied to all series).}

\item{xlim}{(optional) c(numeric, numeric) Gives the x limits (in years) for the graph. Alternatively, you can supply a list to provide different x limits for each panel (not recommended). If unsupplied, a suitable default is chosen (recommended).}

\item{ylim}{(optional) A list of string -> list(min = numeric, max = numeric, nsteps int) pairs. Keys are panel names (e.g. "1", "2", etc). Values are the scale, provided as a list with the keys min, max and nsteps. If unsupplied, a suitable default is chosen (recommended, but will not work well for multipanels).}

\item{legend}{A logical indicating whether to add a legend to the graph (default FALSE).}

\item{legend.ncol}{How many columns do you want the legend to have (if NA, which is the default, arphit will guess for you).}

\item{bar.stacked}{(optional) Logical indicating whether the bar series should be stacked (TRUE, default) or side-by-side (FALSE).}
}
\description{
Quick plot - for quickly creates a single-panel graph. Supports bar and line (and combinations of).
}
\examples{
T <- 24
randomdata <- ts(data.frame(x1 = rnorm(T), x2 = rnorm(T), x3 = rnorm(T, sd = 10),
  x4 = rnorm(T, sd = 5)), start = c(2000,1), frequency = 4)
agg_qplot(randomdata, title = "A Title", subtitle = "A subtitle",
  footnotes = c("a","B"), sources = c("A Source", "Another source"), yunits = "index")

}
\seealso{
\code{vignette("qplot-options", package = "arphit")} for a detailed description of
all the plotting options and how they affect the output.
}
