% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{agg_slides}
\alias{agg_slides}
\title{Make slides in gif}
\usage{
agg_slides(gg_list, filename = NULL, delay = 1, loop = 0)
}
\arguments{
\item{gg_list}{List of gg objects}

\item{filename}{Path to save .gif to. If omitted, .gif will be opened
externally}

\item{delay}{Delay in seconds for each frame}

\item{loop}{Whether to loop through frames. Setting to zero will loop
infinitely (default).}
}
\value{
filename if provided, or NULL if not
}
\description{
Flick through arphitgg objects in a gif
}
