% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{agg_slides_gifski}
\alias{agg_slides_gifski}
\title{Make slides in gif using gifski}
\usage{
agg_slides_gifski(gg_list, filename = NULL, delay = 1, loop = TRUE)
}
\arguments{
\item{gg_list}{List of gg objects}

\item{filename}{Path to save .gif to}

\item{delay}{Delay in seconds for each frame}

\item{loop}{Whether to loop through frames}
}
\value{
filename, if successful
}
\description{
Flick through ggs in a gif
}
