% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{agg_source}
\alias{agg_source}
\title{Add a source (or many sources)}
\usage{
agg_source(source)
}
\arguments{
\item{source}{A string, or vector of strings, to be added as sources}
}
\description{
Add a source (or many sources)
}
\examples{
data  <- data.frame(unemployment = rnorm(20), state = c(rep("A", 10), rep("B", 10)), date = seq.Date(from = as.Date("2017-01-10"), length.out = 10, by = "quarter"))
arphitgg(data) + agg_footnote("Source 1")
arphitgg(data) + agg_footnote(c("Source 1", "Source 2"))

}
\seealso{
\code{vignette("gg-interface", package = "arphit")} for a detailed description of
how to use the ggplot-like interface.
}
