% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{agg_subtitle}
\alias{agg_subtitle}
\title{Add a subtitle or panel subtitle}
\usage{
agg_subtitle(text, panel = NULL)
}
\arguments{
\item{text}{A string for the subtitle.}

\item{panel}{(optional) Specify a panel identifier to add a panel subtitle instead of an overall graph subtitle.}
}
\description{
Add a subtitle or panel subtitle
}
\examples{
data  <- data.frame(unemployment = rnorm(20), state = c(rep("A", 10), rep("B", 10)), date = seq.Date(from = as.Date("2017-01-10"), length.out = 10, by = "quarter"))
arphitgg(data) + agg_subtitle("Graph Subtitle")
arphitgg(data) + agg_subtitle("Panel subtitle", panel = "1")

}
\seealso{
\code{vignette("gg-interface", package = "arphit")} for a detailed description of
how to use the ggplot-like interface.
}
