% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{agg_units}
\alias{agg_units}
\title{Add units}
\usage{
agg_units(units, panel = NULL)
}
\arguments{
\item{units}{A string specifying the units.}

\item{panel}{(optional) Specify a panel identifier to add to a specific panel. If blank, units will be applied to all panels.}
}
\description{
Add units
}
\examples{
arphitgg(data) + agg_units("index")
arphitgg(data) + agg_units("ppt", panel = "1")

}
\seealso{
\code{vignette("gg-interface", package = "arphit")} for a detailed description of
how to use the ggplot-like interface.
}
