% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-constructors.R
\name{agg_vline}
\alias{agg_vline}
\title{Add a vertical line to your graph}
\usage{
agg_vline(x, colour = "black", panel, lwd = 1, lty = 1)
}
\arguments{
\item{x}{The x coordinate to draw the vertical line at}

\item{colour}{The colour of the line (default black)}

\item{panel}{Which panel should the line be placed on? You can specify a vector of panels (e.g. `panel = c("1","3")`) to apply the line to multiple panels at once.}

\item{lwd}{(Optional, default 1) The line width}

\item{lty}{(Optional, default 1) The line type (uses R line types)}
}
\description{
Add a vertical line to your graph
}
\examples{
arphitgg() + agg_vline(x=2003,panel="1")
}
