% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-constructors.R
\name{agg_waterfall}
\alias{agg_waterfall}
\title{Add a waterfall layer to an arphit plot.}
\usage{
agg_waterfall(aes = NULL, data = NULL, colour = c(RBA["Grey5"],
  RBA["Red4"], RBA["DarkGreen3"], RBA["Grey10"]), barcol = NULL,
  panel = "1")
}
\arguments{
\item{aes}{The aesthetic that defines the layer. Will inherit (or parts thereof) if omitted.}

\item{data}{The data to be used. Will inherit from parent if missing.}

\item{colour}{A vector of four colours for (in order) the 'from'/'start' bar,
negative, changes, positive changes and the 'to'/'end' bar}

\item{barcol}{(optional) Outline colours for each bar series (same order as
`colour`)}

\item{panel}{(default = "1") Which panel of the graph to place this layer on. You can specify a vector of panels (e.g. `panel = c("1","3")`) to apply the layer to multiple panels at once.}
}
\description{
Creates a bar chart, showing how the changes from left-most observation to the
right-most observation.
}
\details{
You should ensure that your data are ordered so that your first observation is
the observation you want the waterfall to 'start' from, and your last
observation the one you want the waterfall to 'end' at. Positive and negative
bars are coloured differently to distinguish them, as are the start and end
observations. Sensible defaults are provided, but you can set your own. See
the documentation for the `colour` argument.

This layer should be considered experimental and may undergo breaking
changes.
}
\examples{
data  <- data.frame(x = letters[1:6], y = c(2,1,-0.5,-0.2,0.4,2.7))
arphitgg(data) + agg_waterfall(agg_aes(x=x,y=y,order=x),from="a",to="f")

}
\seealso{
\code{vignette("plotting-options", package = "arphit")} for a detailed description of
all the plotting options
}
