% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-constructors.R
\name{agg_xaxisfreq}
\alias{agg_xaxisfreq}
\title{Specify the frequency of the x ticks for time series graphs}
\usage{
agg_xaxisfreq(freq, panel = NULL)
}
\arguments{
\item{freq}{The frequency, either "decade", "year", "quarter", "month"}

\item{panel}{(optional) Which panel to apply the scale to. If left blank, will be applied to all panels.}
}
\description{
Specify the frequency of the x ticks for time series graphs
}
\seealso{
\code{vignette("plotting-options", package = "arphit")} for a detailed description of
all the plotting options
}
