% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{agg_xaxislabel}
\alias{agg_xaxislabel}
\title{Add an axis label to the x axis}
\usage{
agg_xaxislabel(axislabel, panel = NULL)
}
\arguments{
\item{axislabel}{A string specifying the axis label}

\item{panel}{(optional) Specify a panel identifier to add to a specific panel. If blank, axis label will be applied to all panels.}
}
\description{
Add an axis label to the x axis
}
\examples{
arphitgg(data) + agg_xaxislabel("year")

}
\seealso{
\code{vignette("gg-interface", package = "arphit")} for a detailed description of
how to use the ggplot-like interface.
}
