% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{agg_xlim}
\alias{agg_xlim}
\title{Specify the x limits for a graph}
\usage{
agg_xlim(min, max, panel = NULL)
}
\arguments{
\item{min}{The minimum for the x-axis}

\item{max}{The maximum}

\item{panel}{(optional) Which panel to apply the scale to. If left blank, will be applied to all panels. This is recommended.}
}
\description{
Specify the x limits for a graph
}
\examples{
arphitgg(data) + agg_xlim(min = -10, max = 10)
arphitgg(data) + agg_xlim(min = -10, max = 10, panel = "1")

}
\seealso{
\code{vignette("gg-interface", package = "arphit")} for a detailed description of
how to use the ggplot-like interface.
}
