% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-constructors.R
\name{agg_xunits}
\alias{agg_xunits}
\title{Add units to the x axis (only works for scatter graphs)}
\usage{
agg_xunits(units, panel = NULL)
}
\arguments{
\item{units}{A string specifying the units.}

\item{panel}{(optional) Specify a panel identifier to add to a specific panel. If blank, units will be applied to all panels.}
}
\description{
Add units to the x axis (only works for scatter graphs)
}
\examples{
arphitgg(data) + agg_xunits("index")

}
\seealso{
\code{vignette("plotting-options", package = "arphit")} for a detailed description of
all the plotting options
}
