% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-constructors.R
\name{agg_ylim}
\alias{agg_ylim}
\title{Specify the y scale for a graph}
\usage{
agg_ylim(min, max, nsteps, panel = NULL)
}
\arguments{
\item{min}{The minimum for the y-axis}

\item{max}{The maximum}

\item{nsteps}{The number of steps (includes the minimum and maximum)}

\item{panel}{(optional) Which panel to apply the scale to. If left blank, will be applied to all panels.}
}
\description{
Specify the y scale for a graph
}
\examples{
arphitgg(data, layout = "2b2") + agg_ylim(min = -10, max = 10, nsteps = 5)
arphitgg(data, layout = "2b2") + agg_ylim(min = -10, max = 10, nsteps = 5, panel = "1")

}
\seealso{
\code{vignette("gg-interface", package = "arphit")} for a detailed description of
how to use the ggplot-like interface.
}
