% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arphit.R
\docType{package}
\name{arphit-package}
\alias{arphit}
\alias{arphit-package}
\title{arphit: RBA-style R Plots}
\description{
Easily create RBA-template graphs using R. Timeseries, stacked bar graphs and multipanel.
}
\details{
You'll most likely only need the arphit.tsgraph function. It has many plotting options
to control how your time series plot looks. These are summarised in brief under the 'arguments'
section. Further details about how to use these options and how they affect your plot
can be found in the vignette.
}
