% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-constructors.R
\name{agg_abline}
\alias{agg_abline}
\title{Add an AB line to your graph}
\usage{
agg_abline(x = NULL, y = NULL, x1 = NULL, y1 = NULL, x2 = NULL,
  y2 = NULL, color = NULL, panel, lwd = 1, lty = 1)
}
\arguments{
\item{x}{Draw a vertical line at x (omit to draw a specific AB line)}

\item{y}{Draw a horizontal at y omit to draw a specific AB line)}

\item{x1}{For specific AB lines: the first x coordinate}

\item{y1}{For specific AB lines: the first y coordinate}

\item{x2}{For specific AB lines: the second x coordinate}

\item{y2}{For specific AB lines: the second y coordinate}

\item{color}{(optional) The color of the AB line (default black)}

\item{panel}{Which panel should the line be placed on?}

\item{lwd}{(Optional, default 1) The linewidth}

\item{lty}{(Optional, default 1) The line type}
}
\description{
You need specify only one of x, or y, or x1,x2,y1,y2
}
\examples{
arphitgg(data) + agg_abline(x = 2001, color = RBA["Blue1"], panel = "1") +
  agg_abline(y = -0.5, color = RBA["Red1"], panel = "1")

arphitgg(data) + agg_abline(x1 = 2000, y1 = -0.1, x2 = 2002, y2 = 0.5, color = RBA["Blue1"], panel = "1")

}
\seealso{
\code{vignette("plotting-options", package = "arphit")} for a detailed description of
all the plotting options
}
