% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-constructors.R
\name{agg_aes}
\alias{agg_aes}
\title{Define an aesthetic for a graph, or a graph layer.}
\usage{
agg_aes(x, y, group = NULL, facet = NULL)
}
\arguments{
\item{x}{Which series is the x variable.}

\item{y}{Which series are you plotting on the y axis.}

\item{group}{If your data are in long form, which variable defines the groups.}

\item{facet}{If you data are in long form, which variable defines the facets (facets split data across panels, see the gg-interface vignette for example)}
}
\description{
If specified as part of a agg_line or agg_col, fields left blank will be inherited from the parent.
}
\seealso{
\code{vignette("plotting-options", package = "arphit")} for a detailed description of
all the plotting options
}
