% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-constructors.R
\name{agg_arrow}
\alias{agg_arrow}
\title{Add an arrow}
\usage{
agg_arrow(tail.x, tail.y, head.x, head.y, color = "black", panel,
  lwd = 1)
}
\arguments{
\item{tail.x}{The x coordinate of the arrow tail}

\item{tail.y}{The y coordinate of the arrow tail}

\item{head.x}{The x coordinate of the arrow head}

\item{head.y}{The y coordinate of the arrow head}

\item{color}{The color of the arrow}

\item{panel}{Which panel should the arrow be placed on? You can specify a vector of panels (e.g. `panel = c("1","3")`) to apply the arrow to multiple panels at once.}

\item{lwd}{(Optional, default 1) The linewidth of the arrow}
}
\description{
Add an arrow
}
\examples{
arphitgg(data) +
  agg_arrow(tail.x = 2002, tail.y = 0, head.x = 2003, head.y = 1,
            color = RBA["Blue1"], panel = "1")

}
\seealso{
\code{vignette("plotting-options", package = "arphit")} for a detailed description of
all the plotting options
}
