% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-constructors.R
\name{agg_autolabel}
\alias{agg_autolabel}
\title{Add automatically placed label}
\usage{
agg_autolabel(quiet = FALSE, arrow_lines = TRUE, arrow_bars = FALSE)
}
\arguments{
\item{quiet}{(logical, default FALSE) Do not output progress of autolabeller}

\item{arrow_lines}{(logical, default TRUE) Add arrows between labels and line series
if there isn't line of sight between the label and the series.}

\item{arrow_bars}{(logical, default FALSE) Add arrows between labels and bar series
if there isn't line of sight between the label and the series.}
}
\description{
Add automatically placed label
}
\examples{
data <- data.frame(x=1:10,y1=rnorm(10),y2=rnorm(10))
arphitgg(data,agg_aes(x=x)) +
  agg_line(agg_aes(y=y1)) +
  agg_line(agg_aes(y=y2)) +
  agg_autolabel()

}
\seealso{
\code{vignette("plotting-options", package = "arphit")} for a detailed description of
all the plotting options
}
