% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-constructors.R
\name{agg_bgshading}
\alias{agg_bgshading}
\title{Add background shading}
\usage{
agg_bgshading(x1 = NA, y1 = NA, x2 = NA, y2 = NA,
  color = RBA["Grey2"], panel)
}
\arguments{
\item{x1}{The bottom left x coordinate (omit to have the shading automatically snap to the edge of the panel)}

\item{y1}{The bottom left y coordinate (omit to have the shading automatically snap to the edge of the panel)}

\item{x2}{The top right x coordinate (omit to have the shading automatically snap to the edge of the panel)}

\item{y2}{The top right y coordinate (omit to have the shading automatically snap to the edge of the panel)}

\item{color}{(optional) The color of the AB line (default grey)}

\item{panel}{Which panel should the background shading be placed on?}
}
\description{
Add background shading
}
\examples{
arphitgg(data) + agg_bgshading(x1 = 2001, x2 = 2002, panel = "1")
arphitgg(data) + agg_bgshading(y1 = 0.5, y2 = -0.5, panel = "1")

}
\seealso{
\code{vignette("plotting-options", package = "arphit")} for a detailed description of
all the plotting options
}
