% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-operators.R
\name{agg_draw}
\alias{agg_draw}
\title{Draw a defined graph}
\usage{
agg_draw(gg, filename = NULL)
}
\arguments{
\item{gg}{An arphitgg built graph.}

\item{filename}{(optional) If specified, save image to filename instead of displaying in R. Supports png, pdf, emf, emf+, svg and xlsx extensions.}
}
\description{
Draw a defined graph
}
\seealso{
\code{vignette("plotting-options", package = "arphit")} for a detailed description of
all the plotting options
}
