% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{agg_footnote}
\alias{agg_footnote}
\title{Add a footnote (or many footnotes)}
\usage{
agg_footnote(footnote)
}
\arguments{
\item{footnote}{A string, or vector of strings, to be added as footnotes.}
}
\description{
Add a footnote (or many footnotes)
}
\examples{
data  <- data.frame(unemployment = rnorm(20), state = c(rep("A", 10), rep("B", 10)), date = seq.Date(from = as.Date("2017-01-10"), length.out = 10, by = "quarter"))
arphitgg(data) + agg_footnote("Here is a footnote")
arphitgg(data) + agg_footnote(c("Here is a footnote", "And a second one"))

}
\seealso{
\code{vignette("gg-interface", package = "arphit")} for a detailed description of
how to use the ggplot-like interface.
}
