% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-constructors.R
\name{agg_label}
\alias{agg_label}
\title{Add a label}
\usage{
agg_label(text, x, y, panel, colour = "black", size = 20, color)
}
\arguments{
\item{text}{The text to display on your plot}

\item{x}{The x coordinate of the center of your label}

\item{y}{The y coordinate of the center of your label}

\item{panel}{Which panel should the label be placed on? You can specify a
vector of panels (e.g. `panel = c("1","3")`) to apply the label to multiple
panels at once.}

\item{colour}{The colour of your text}

\item{size}{Font size (default 20)}

\item{color}{(Deprecated; use colour instead) The colour of your text}
}
\description{
Add a label
}
\examples{
arphitgg() + agg_label("Here is a label", 2003, 0.2, "1", RBA["Red3"])

}
\seealso{
\code{vignette("plotting-options", package = "arphit")} for a
detailed description of all the plotting options
}
