% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-constructors.R
\name{agg_legend}
\alias{agg_legend}
\title{Add a legend to the graph}
\usage{
agg_legend(ncol = NULL)
}
\arguments{
\item{ncol}{(optional) Specify the number of columns in the legend (if left blank, arphit will guess)}
}
\description{
Add a legend to the graph
}
\examples{
data <- data.frame(x = 1:10, y = 1:10)
arphitgg(data) + agg_line(agg_aes(x = x, y = y)) + agg_legend()

}
\seealso{
\code{vignette("plotting-options", package = "arphit")} for a detailed description of
all the plotting options
}
