% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot.R
\name{agg_qplot}
\alias{agg_qplot}
\title{RBA-style graphs in R}
\usage{
agg_qplot(data, series = NULL, x = NULL, layout = "1", bars = NULL,
  filename = NULL, title = NULL, subtitle = NULL, paneltitles = NULL,
  panelsubtitles = NULL, footnotes = NULL, sources = NULL,
  yunits = NULL, col = NULL, pch = NULL, lty = NULL, lwd = NULL,
  barcol = NULL, xlim = NULL, ylim = NULL, legend = FALSE,
  legend.ncol = NA, portrait = FALSE, bar.stacked = TRUE,
  dropxlabel = FALSE, joined = TRUE, srt = 0, showallxlabels = NULL)
}
\arguments{
\item{data}{Object containing the series you want to plot. Can be a data.frame, tibble or ts. Can also be a list of the above, with separate for each panel.}

\item{series}{(optional) A list of vectors of string names of the series from data you wish to plot. The keys in the list must be "1", "2", etc to indicate which panel the series should be plotted on.
Panel numbers are sequential left-to-right, top-to-bottom. For one-panel and top-and-bottom two panel, the right axis is counted as a separate panel. Thus, for a top-and-bottom, panels "1" and "2" are the top panel, left and right respectively; likewise "3" and "4" for the bottom panel.
You do not need to supply all panels. For instance, in a one-panel, there is no need to supply series for panel "2" if you want all series on the left hand side axis.
Alternatively if just a vector of string names is supplied, it will be assumed all series are being plotted in panel 1. Similarly, if series is not supplied at all, all series will be plotted in panel 1.}

\item{x}{(optional) A list specifying x variables for each of your panels. ts objects passed in as data will automatically use dates.}

\item{layout}{(optional) A string indicating the layout of the chart. Valid options are "1" (single panel), "2v" (side-by-side two panel), "2h" (top and bottom two panel), "2b2" (two-by-two four panel chart), "3v" (side-by-side three panel), "3h" (horizonal three panel), "3b2" (three-by-two six panel chart), "4h" (horizonal four panel) and "4b2" (four-by-two eight panel). Defaults to single panel if not supplied.}

\item{bars}{(optional) Vector of string names indicating which series should be bars, rather than lines. Alternatively, if you set `bars = TRUE` all series will plot as bars.}

\item{filename}{(optional) If specified, save image to filename instead of displaying in R. Supports pdf, emf and png extensions.}

\item{title}{(optional) A string indicating the title for the entire chart. Passing NULL (or omitting the argument) will suppress printing of title.}

\item{subtitle}{(optional) A string indicating the subtitle for the entire chart. Passing NULL (or omitting the argument) will suppress printing of subtitle.}

\item{paneltitles}{(optional) A list of string -> string pairs indicating panel titles. Keys must be "1", "2", etc to indicate which panel the title is for.}

\item{panelsubtitles}{(optional) A list string -> string pairs indicating panel titles. See paneltitles.}

\item{footnotes}{(optional) A vector strings, corresponding to the footnotes, in order.}

\item{sources}{(optional) A vector of strings, one entry for each source.}

\item{yunits}{(optional) A list of string -> string pairs indicating the units to be used on each panel (/axes, see notes to series for explanation on how right-hand-side axes are treated). Alternatively, providing just a string will assign that to all panels. If not supplied, a per cent sign will be used.}

\item{col}{(optional) A list of string -> misc pairs. The keys should be series names, and the values colours for each series (any colour accepted by R is fine.) You need not supply colours for all series. Default colours will be assigned  (cycling through) to series without assigned colours. Alternatively, you can supply a single value to apply to all series.}

\item{pch}{(optional) A list of string -> int pairs. The keys should be series names, and the values markers (pch values) for each series. Defaults to none (NA). Can be supplied as list, or a single value (which will be applied to all series).}

\item{lty}{(optional) A list of string -> int pairs. The keys should be series names, and the values line types (lty values) for each series. Defaults to solid (1). Can be supplied as list, or a single value (which will be applied to all series).}

\item{lwd}{(optional) A list of string -> numeric pairs. The keys should be series names, and the values line width, relative to default, for each series. Defaults to 1. Can be supplied as list, or a single value (which will be applied to all series).}

\item{barcol}{(optional) A list of string -> misc pairs.  The keys should be series names, and the values outline colours for each bar series (any colour accepted by R is fine.) You need not supply colours for all series. Bar series without assigned colours are given no outline by default. Alternatively, you can supply a single value to apply to all series.}

\item{xlim}{(optional) c(numeric, numeric) Gives the x limits (in years) for the graph. Alternatively, you can supply a list to provide different x limits for each panel (not recommended). If unsupplied, a suitable default is chosen (recommended).}

\item{ylim}{(optional) A list of string -> list(min = numeric, max = numeric, nsteps int) pairs. Keys are panel names (e.g. "1", "2", etc). Values are the scale, provided as a list with the keys min, max and nsteps. If unsupplied, a suitable default is chosen (recommended, but will not work well for multipanels).}

\item{legend}{A logical indicating whether to add a legend to the graph (default FALSE).}

\item{legend.ncol}{How many columns do you want the legend to have (if NA, which is the default, arphit will guess for you).}

\item{portrait}{(optional) Logical indicating whether the layout should be a landscape size (FALSE, default), or a taller portrait size (TRUE).}

\item{bar.stacked}{(optional) Logical indicating whether the bar series should be stacked (TRUE, default) or side-by-side (FALSE).}

\item{dropxlabel}{(optional) Logical indicating whether the first xlabel of right hand panels in 2v , 3v, 2b2, 3b2, and 4b2 should be ignored (prevents overlapping of last xlabel on left panel with first on right). FALSE by default.}

\item{joined}{(optiona) Logical indicating whether you want to join between missing observations (TRUE()), or break the series (FALSE). TRUE by default.}

\item{srt}{(default 0) Orientation adjustment for xlabels. In degrees; 0 is horizontal.}

\item{showallxlabels}{(optional) (Only for categorical graphs) Force all x labels to show? By default, this is false for numeric categorical and true for non-numeric categorical.}
}
\description{
Quickly creates a (potentially multipanel) graph. Supports bar and line (and combinations of).
}
\examples{
T <- 24
randomdata <- ts(data.frame(x1 = rnorm(T), x2 = rnorm(T), x3 = rnorm(T, sd = 10),
  x4 = rnorm(T, sd = 5)), start = c(2000,1), frequency = 4)
agg_qplot(randomdata, series = list("1" = c("x1", "x2"), "2" = c("x3", "x4")),
  bars = c("x3","x4"), layout = "2v", title = "A Title", subtitle = "A subtitle",
  footnotes = c("a","B"), sources = c("A Source", "Another source"), yunits = "index")

}
\seealso{
\code{vignette("qplot-options", package = "arphit")} for a detailed description of
all the plotting options and how they affect the output.
}
