% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-constructors.R
\name{agg_rename_series}
\alias{agg_rename_series}
\title{Rename series in your plot (for legends and autolabeller)}
\usage{
agg_rename_series(mapping = list(), panel = NULL)
}
\arguments{
\item{mapping}{A list with values the existing series names and names the new series names}

\item{panel}{(optional) Only apply mapping to specific panels}
}
\description{
Rename series in your plot (for legends and autolabeller)
}
\examples{
arphitgg(data.frame(x=1:10,y=1:10)) + agg_line() +
  agg_rename_series(list('Nicer series name' = 'y'))

}
