% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-constructors.R
\name{agg_source}
\alias{agg_source}
\title{Add a source (or many sources)}
\usage{
agg_source(source)
}
\arguments{
\item{source}{A string, or vector of strings, to be added as sources}
}
\description{
Add a source (or many sources)
}
\examples{
arphitgg(data) + agg_source("Source 1")
arphitgg(data) + agg_source(c("Source 1", "Source 2"))

}
\seealso{
\code{vignette("gg-interface", package = "arphit")} for a detailed description of
how to use the ggplot-like interface.
}
