% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{agg_title}
\alias{agg_title}
\title{Add a title or panel title}
\usage{
agg_title(text, panel = NULL)
}
\arguments{
\item{text}{A string for the title.}

\item{panel}{(optional) Specify a panel identifier to add a panel title instead of an overall graph title.}
}
\description{
Add a title or panel title
}
\examples{
data  <- data.frame(unemployment = rnorm(20), state = c(rep("A", 10), rep("B", 10)), date = seq.Date(from = as.Date("2017-01-10"), length.out = 10, by = "quarter"))
arphitgg(data) + agg_title("Graph Title")
arphitgg(data) + agg_title("Panel title", panel = "1")

}
\seealso{
\code{vignette("gg-interface", package = "arphit")} for a detailed description of
how to use the ggplot-like interface.
}
