% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-constructors.R
\name{agg_title}
\alias{agg_title}
\title{Add a title or panel title}
\usage{
agg_title(text, panel = NULL)
}
\arguments{
\item{text}{A string for the title.}

\item{panel}{(optional) Specify a panel identifier to add a panel title instead of an overall graph title.  You can specify a vector of panels (e.g. `panel = c("1","3")`) to apply the title to multiple panels at once.}
}
\description{
Add a title or panel title
}
\examples{
arphitgg(data) + agg_title("Graph Title")
arphitgg(data) + agg_title("Panel title", panel = "1")

}
\seealso{
\code{vignette("plotting-options", package = "arphit")} for a detailed description of
all the plotting options
}
