% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{arphitgg}
\alias{arphitgg}
\title{Create an arphit graph to be built using the ggplot-like interface.}
\usage{
arphitgg(data = NULL, aes = NULL, layout = "1", portrait = FALSE,
  dropxlabel = FALSE)
}
\arguments{
\item{data}{(Optional) Data to be used for the plot. Can be left blank, but must then be supplied for each layer.}

\item{aes}{(Optional) The aesthetic that defines your graph. Can be left blank, but must then be supplied for each layer. Layers that don't specify aesthetics will inherit missing parts of aesthetic from here.}

\item{layout}{(default = "1")}

\item{portrait}{(default = false) Logical indicating whether the layout should be a landscape size (FALSE, default), or a taller portrait size (TRUE).}

\item{dropxlabel}{(optional) Logical indicating whether the first xlabel of right hand panels in 2v and 2b2 should be ignored (prevents overlapping of last xlabel on left panel with first on right). FALSE by default.}
}
\description{
Create an arphit graph to be built using the ggplot-like interface.
}
\seealso{
\code{vignette("gg-interface", package = "arphit")} for a detailed description of
how to use the ggplot-like interface.
}
