% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-constructors.R
\name{arphitgg}
\alias{arphitgg}
\title{Create an arphit graph to be built using the ggplot-like interface.}
\usage{
arphitgg(data = NULL, aes = NULL, layout = "1", portrait = FALSE,
  dropxlabel = FALSE, srt = 0, showallxlabels = NULL, joined = TRUE,
  plotsize = LANDSCAPESIZE, log_scale = "")
}
\arguments{
\item{data}{(Optional) Data to be used for the plot. Can be left blank, but must then be supplied for each layer.}

\item{aes}{(Optional) The aesthetic that defines your graph. Can be left blank, but must then be supplied for each layer. Layers that don't specify aesthetics will inherit missing parts of aesthetic from here.}

\item{layout}{(default = "1") The layout of the graph. Valid options are "1", "2v", "2h", "2b2", "3v", "3h", "3b2", "4h", "4b2".}

\item{portrait}{(default = FALSE) Logical indicating whether the layout should be a landscape size (FALSE, default), or a taller portrait size (TRUE).}

\item{dropxlabel}{(optional) Logical indicating whether the first xlabel of right hand panels in 2v and 2b2 should be ignored (prevents overlapping of last xlabel on left panel with first on right). FALSE by default.}

\item{srt}{(default 0) Orientation adjustment for xlabels. In degrees; 0 is horizontal.}

\item{showallxlabels}{(optional) (Only for categorical graphs) Force all x labels to show? By default, this is false for numeric categorical and true for non-numeric categorical.}

\item{joined}{(default TRUE) Logical indicating whether you want to join between missing observations (TRUE()), or break the series (FALSE).}

\item{plotsize}{(optional) A vector of two variables specifying the height and width of your graph, respectively. (Default 5.53, 7.5). Ignored if portrait is set to TRUE}

\item{log_scale}{(optional) Set one or both of y and x axes as log scales? Use "y" for just y axis, "x" for just x and "xy" for both}
}
\description{
Create an arphit graph to be built using the ggplot-like interface.
}
\seealso{
\code{vignette("plotting-options", package = "arphit")} for a detailed description of
all the plotting options
}
