% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-operators.R
\name{+.arphit.gg}
\alias{+.arphit.gg}
\title{Add a layer or element to an arphitgg graph.}
\usage{
\method{+}{arphit.gg}(gg, element)
}
\arguments{
\item{gg}{An arphitgg built graph.}

\item{element}{The element to add to the graph.}
}
\description{
Add a layer or element to an arphitgg graph.
}
\seealso{
\code{vignette("plotting-options", package = "arphit")} for a detailed description of
all the plotting options
}
