% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{print.arphit.gg}
\alias{print.arphit.gg}
\title{Draw a defined graph}
\usage{
\method{print}{arphit.gg}(x, ...)
}
\arguments{
\item{x}{An arphitgg built graph.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Draw a defined graph
}
\seealso{
\code{vignette("gg-interface", package = "arphit")} for a detailed description of
how to use the ggplot-like interface.
}
