% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colpal.R
\docType{data}
\name{RBA}
\alias{RBA}
\title{RBA colour palette}
\format{An object of class \code{character} of length 108.}
\usage{
RBA
}
\description{
A list of RBA-standard colours to use. Use them by using RBA["Red1"] in a colour declaration. For instance:
}
\details{
arphit.tsgraph(data, col = list("x2" = RBA["Red1"], "x4" = RBA["Green3"]))

To see a list of all the available colours, type `vignette("rbacolours")`
}
\keyword{datasets}
